fixed segfault on 64 bit platforms

+++ deadbeef.h
@@ -126,8 +126,8 @@ typedef struct {
     const char *(*get_config_dir) (void);
     void (*quit) (void);
     // threading
-    int (*thread_start) (void (*fn)(uintptr_t ctx), uintptr_t ctx);
-    int (*thread_join) (int tid);
+    intptr_t (*thread_start) (void (*fn)(uintptr_t ctx), uintptr_t ctx);
+    int (*thread_join) (intptr_t tid);
     uintptr_t (*mutex_create) (void);
     void (*mutex_free) (uintptr_t mtx);
     int (*mutex_lock) (uintptr_t mtx);
